Небольшие правки

+++ src/main/java/com/nb/globalerp/training/sitebackendglobalerp/api/dto/response/GroupMemberResponse.java
@@ -1,10 +1,8 @@
 package com.nb.globalerp.training.sitebackendglobalerp.api.dto.response;
 
-import java.math.BigDecimal;
-
 public record GroupMemberResponse(
         int id,
         Integer studentId,
         Integer groupId,
-        BigDecimal completionPercent
+        Float completionPercent
 ) {}
